testsuite: skip tests requiring xmds1 if xmds1 isn't installed
This cuts down on the noise when testing and makes my life
easier. Note that the skip functionality in the unittest
class is only available in python 2.7 and later; the code
checks for the python version before trying to skip.

Technically XMDS2 is written for earlier versions of python as
well, so developers using the testing framework who have an
earlier version of python won't get the benefit of this. As far
as I'm aware this is a set of measure zero.

If we want to include them we just need to replace "unittest"
with "unittest2" everywhere in run_tests.py, as well as pulling
in unittest2 in our installers via "sudo easy_install unittest2".

I've tested this and it works, but I didn't want to make the
installers pull in even more stuff, so I haven't implemented it.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2673 41b19578-b77a-4e3b-9682-c8570d237623