Bug fix: Constant IP or EX operators with dependencies don't work with multi-path drivers.
The problem was that the IP or EX operators were being initialised before their dependencies, and as an optimisation, they were only initialised once per simulation (not once per path).  However, the vectors that they depended on were only initialised later in the simulation, so when the IP or EX operators were initialised, the dependencies would be zero.

This commit fixes the bug by disabling the optimisation in this case.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2674 41b19578-b77a-4e3b-9682-c8570d237623