Simplify examples in the documentation
Remove the  <?xml ...> elements from the worked examples in
the documentation, except for the very first one where its
purpose is explained. Also remove the format and filename tags
from the <output> element.

This means the doc examples focus on what they're trying to
demonstrate with a minimum of extraneous fluff.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2687 41b19578-b77a-4e3b-9682-c8570d237623