Linux installer: provide fallback option if fftw version not found
Get fftw-3.3.2 from the main download area of www.fftw.org. If that
file isn't present (presumably because a new version has been released)
get fftw-3.3.1 from their FTP site instead. This site contains all
non-current versions, so we can be certain of getting the file.

We go for 3.3.1 on the FTP site in case there's a lag between taking
3.3.2 off the main download area and adding it to the FTP site.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2691 41b19578-b77a-4e3b-9682-c8570d237623