You can now initialise a vector by integrating over dimensions of other vectors.
XMDS's syntax has permitted integrating over a dimension when initialising a vector. E.g. integrating over a 1d vector in the construction of a 0d vector.

However, this was a flexibility that was not intended.  Of course, Sebastian tried to use it and was given an unhelpful internal error message for his efforts.

This commit adds support for this feature, and adds a test case.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2700 41b19578-b77a-4e3b-9682-c8570d237623