Fix bug with high-dimensional distributed-mpi simulations where the sampling basis order is unusual.
If you had three dimensions, "x y z", and the MPI dimensions were "x" and "y", if you specified a sampling basis of "z x y", XMDS would actually try to keep the "z" dimension first.  This was incorrect and would lead to a simulation run-time error as XMDS wouldn't have a transform plan to be able to transform to that basis.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2706 41b19578-b77a-4e3b-9682-c8570d237623