Reduce by one the number of FFTs needed to evaluate the EX operator in some cases.
The evaluation of the EX operator uses an additional FFT because it transforms the solution to spectral space to evaluate the EX operator than inverse transforms *both* the solution and the EX operator.  As the solution is never modified, there's no need to do the inverse transform.

The solution I've implemented is to copy the solution vector to the results vector of the EX operator (provided this is safe) and then transform this vector.  As the original vector is never transformed, it doesn't need to be inverse transformed.  This optimisation is only safe in the common, but restricted situation in which only one EX operator is used (e.g. L[u], not Lx[u] and Ly[u]), and the EX operator is applied to all components of the vector.  E.g. if there's u and uDot, then just computing Lxx[u] won't get the optimisation.

Add testcases.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2744 41b19578-b77a-4e3b-9682-c8570d237623