Sometimes you really do need access to an arbitrary grid point, and the existing nonlocal dimension access just isn't enough.
Previously, in these situations you had to resort to writing out the access to the underlying array by hand, knowing all of XMDS's internal variables and memory layouts.  Now, for these hacks you can just use the syntax:

psi(x_index => 3)

This will give you access the the grid point with index 3.  Note this is the *fourth* grid point because the x_index in zero-based.  The format for the variable is just <dimrep-name>_index.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2780 41b19578-b77a-4e3b-9682-c8570d237623