Significant speedups to the nonlocal dimension access parser.
The nonlocal dimension access parser was taking over a minute, sometimes much, much more on simulations with ~200 variables.

The parser is now much faster for such simulations, and even skipped entirely in the common case that the string "=>" doesn't appear in the code, so there can't be any nonlocal dimension access.

Added test case derived from the script that caused this problem in the first case.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2798 41b19578-b77a-4e3b-9682-c8570d237623