Make external entity references actually work.
External entity references are a standard XML feature that allows you to include the content of another XML document in the current document.

For example, include the following at the top of your document:
<!DOCTYPE simulation [
<!ENTITY SimulationCode SYSTEM "lorenz_code.txt" >
]>

And you can now use "&SimulationCode;" to refer to the content of lorenz_code.txt in your XMDS script.
NOTE: &SimulationCode; must not appear in a CDATA block.  To include equations of motion, put the <![CDATA[ ]]> declaration around your equations of motion in lorenz_code.txt

See the example lorenz_entity.xmds for a demonstration of this feature.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2804 41b19578-b77a-4e3b-9682-c8570d237623