Document the "home_space" integrator attribute that I had forgotten about.
"home_space" specifies which space time-stepping will occur in.  By default it is position space, but if your dynamics are diagonal in Fourier space, then if your time-stepping is done in Fourier space too then you can integrate your equations with zero Fourier transforms (with the possible exception of one at the start to get everything into the Fourier basis)!

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2813 41b19578-b77a-4e3b-9682-c8570d237623