Created a special mode and a special output function for DEGS
I have added a special degs mode triggered by the command line option '-d' or '--degs'.
If there is a parsing error and if xmds was launched in degs mode it calls a separate function that formats the error output for DEGS. It leaves the current implementation of the output unchanged and it leaves you free to modify the output format as you wish.

Author: Jacopo Sabbatini <dedalusj@gmail.com>

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2827 41b19578-b77a-4e3b-9682-c8570d237623