Fix compatibility with gcc-4.8
Previously I had a hack to rename std::complex to std::Complex so that we could use 'complex' as the type that we wanted.  gcc 4.8 now has #undef complex at the start of the C++ "complex" include, so that doesn't work any more.  I've now done the opposite.  All uses of 'complex' in XMDS simulations are now #defined to XMDSComplexType which is the new complex type.

The short of it is that everything will work as normal, but error messages may be slightly different.  They may refer to "XMDSComplexType" where they previously referred to "complex".

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2829 41b19578-b77a-4e3b-9682-c8570d237623