Fix a bug in the Hermite Gauss Fourier basis.
The bug was that 'kx' for the Hermite Gauss basis was effectively negated, so that 'kx' was replaced with '-kx'.  If you were computing the kinetic energy operator, you wouldn't have noticed this.  But if you were computing quantities like kx*psi, then this bug would have affected you.

The 'nx' <--> 'kx' transforms (and 'x' <--> 'kx' transforms) are not as efficient as they could be for the Hermite-Gauss basis.  Most people will only be using the 'x', 'nx', and perhaps the 'x_4f' bases which are reasonably efficient.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2835 41b19578-b77a-4e3b-9682-c8570d237623