Stochastic error checking with runtime noise generation now works.
Previously, the noise seeds would be generated at the beginning of the full step and the half step.  The noises for the half step would therefore not be the same as for the full step.  This fixes that problem.

As this is was a bug with stochastic simulations, I'm unsure how best to add test cases to check for this.  The existing test suite still passes with this change…  Ideas anyone?

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2858 41b19578-b77a-4e3b-9682-c8570d237623