Fix a situation where an unhelpful error would be presented to the user if a vector needed to be integrated over when initialising a vector.
Now, the error is avoided, and when a similar error is generated in the future, no line number information is provided (as it would be incorrect), and the user is asked to report a bug to the developers.  This is a sensible approach because this is an internal error anyway.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2860 41b19578-b77a-4e3b-9682-c8570d237623