The "constant" attribute to IP & EX operators is now optional and should be considered only for advanced use cases.
The "constant" attribute continues to work as before, but now there are sensible defaults so that you can omit it.  All examples now have the "constant" attribute removed and it is now only mentioned in the IP/EX operator section of the user documentation.

Please note that the default for the EX operator is constant="no", not constant="yes" as one might expect.  Typically the EX operator is very cheap to compute, and on modern computers, a complex multiplication is much faster than a memory reference, so we're better off re-computing this every time.  Of course, if you have an expensive EX operator you can still use constant="yes".

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2884 41b19578-b77a-4e3b-9682-c8570d237623