Optimise NonConstantIPOperator to not recompute IP operators when the last time step is the same as the new one.
Not all propagation step fractions need to be kept in memory at once, and currently the NonConstantIPOperator is optimised to reduce this memory usage.  However, now that we can re-use past results, it will probably be worthwhile using the extra memory to save recomputing them.

As an example, the ARK45 integrator has 5 propagation step fractions, however only 2 of them are needed simultaneously, so only enough memory is allocated for 2 of them.  If we allocate memory for all 5 (and use all 5) we should see a performance improvement on adaptive integrators with IP operators.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2887 41b19578-b77a-4e3b-9682-c8570d237623