All IP operators for adaptive integrators are now much faster, but use more memory.
In particular, this is likely to significantly speed up 1D simulations.  It will have a similar impact on 2D simulations that haven't separated their IP operators for each dimension, but a much smaller impact on those that have.

The extra memory requirements are unlikely to be an issue in 1D, and in higher dimensions it will only be an issue if you don't use separate IP operators for separate dimensions (which you should, because it's faster).

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2888 41b19578-b77a-4e3b-9682-c8570d237623