The Bulirsch-Stoer Algorithm (Richardson extrapolation).
This code was written by Sean Wild's computer science team, which included Damien Beard and one or two others...

This patch introduces a *fixed-order* Bulirish-Stoer algorithm which can be used by choosing the "BS" algorithm.  You specify the number of extrapolations with the "extrapolations" attribute on the integrator.  For the "BS" integrator, the order of the algorithm should be twice the number of the "extrapolations" parameter.

Although this code is usable in its present form, it's incomplete as the aim was an adaptive algorithm which is simultaneously adaptive-order and adaptive-step.  The present code is fixed-order and fixed-step, however it should serve as a good foundation for doing the full adaptive-order adaptive-step case.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2899 41b19578-b77a-4e3b-9682-c8570d237623