Fix the Bessel-Neumann transform for order > 0
The previous version of the bessel-neumann transform only worked for order 0 (which is all anyone will probably ever want).  This commit fixes this.

For reasons I don't understand yet, the error of the bessel-neumann transform seems to be about 5 orders of magnitude larger than bessel.  If I take something of order 1 perform the transform and then inverse transform, then I get an absolute error of order 10^-15 if I use the bessel transform, but 10^-10 if I use the bessel-neumann transform.  10^-10 is almost always going to be small enough, but it's still odd.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2915 41b19578-b77a-4e3b-9682-c8570d237623