Don't require Sphinx as a dependency
This will reduced the size of the download when installing
XMDS2 from Ubuntu or Debian repositories.

The rationale is that we ship the documentation in the
tarball, so the end user doesn't need to use Sphinx to
build the documentation. If the end user installs using
an installer rather than the tarball, Sphinx will be
installed.

If actually install from source, and want to build the 
documentation, they'll have to install Sphinx themselves.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2921 41b19578-b77a-4e3b-9682-c8570d237623