Reduce size of release tarball
There are a lot of files and directories the end
(non-developer) user doesn't need, so they don't
need to be in the tarball. This patch alters the
script that generates the tarball so now these files
are omitted.

At the moment the python files that are generated
from the Cheetah templates are still included in the
tarball, so that's something that could be omitted
in future, provided setup.py adds an install hook to
run "make" and generate them.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2924 41b19578-b77a-4e3b-9682-c8570d237623