Revert commit 2927
Moving the "make" command to compile the Cheetah templates into
setup.py turned out to be a bad plan. Since setup.py is run as
root, we end up with a bunch of compiled files plus various other
files (such as Version.py) owned by root when they shouldn't be.

This causes problems when running "make update" as an unprivileged
user.


git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2930 41b19578-b77a-4e3b-9682-c8570d237623