Almost entirely revert the HDF5 patch, and fix correctly... I think.
using conf.parse_flags was never going to work for static libraries because the static libraries were ending up in the wrong part of the compile line, *before* the xmds script's code, and not after.  This is because parse_flags doesn't completely understand h5cc's output, and h5cc isn't as smart as waf.

git-svn-id: svn+ssh://svn.code.sf.net/p/xmds/code/trunk/xpdeint@2964 41b19578-b77a-4e3b-9682-c8570d237623